requirements.txt
bittensor>=6.0.0
torch>=2.0.0
numpy>=1.24.0
rich>=13.0.0
aiohttp>=3.8.0
requests>=2.31.0
python-dotenv>=1.1.0